requirements.txt
cv2
skimage.segmentation
scipy
matplotlib
json